Removed Internals.Headers
