+++ Sources/RequestDL/Internals/Sources/Session/Session/Models/Internals.ClientResponseReceiver.swift
@@ -91,7 +91,7 @@ extension Internals {
                         minor: head.version.minor,
                         major: head.version.major
                     ),
-                    headers: Headers(head.headers),
+                    headers: .init(head.headers),
                     isKeepAlive: head.isKeepAlive
                 )
