+++ Sources/RequestDL/Properties/Sources/Payloads/Payload/Models/PayloadNode.swift
@@ -39,7 +39,7 @@ struct PayloadNode: PropertyNode {
 
     // MARK: - Private methods
 
-    private func isURLEncoded(_ headers: Internals.Headers) -> Bool {
+    private func isURLEncoded(_ headers: HTTPHeaders) -> Bool {
         headers.contains(name: "Content-Type") {
             $0.range(of: "x-www-form-urlencoded", options: .caseInsensitive) != nil
         }
